Revert c9029c6..4be228b
This rolls back to commit c9029c631007ce043ebd1c7126d663f97e6d1851.

+++ src/PhotoScrollViewController.h
@@ -38,7 +38,7 @@
     
 }
 
-@property int currentImageId;
+@property long currentImageId;
 @property BOOL islastpageList;
 @property bool ishidebar;
 @property BOOL isPlaying;    //播放状态
